Extract and document path manipulation utilities
