+++ src/utils.ts
@@ -21,26 +21,7 @@
 // SOFTWARE.
 
 import * as _ from 'lodash';
-
-export function normalizePath(path: string): string {
-  if (!path) {
-    return '';
-  }
-  return path.replace(/^\//, '').replace(/\/$/, '');
-}
-
-export function pathParts(path: string): string[] {
-  if (!path || path === '' || path === '/') {
-    return [];
-  }
-  return normalizePath(path).split('/');
-}
-
-export function joinPath(base: string, child: string) {
-  return pathParts(base)
-    .concat(pathParts(child))
-    .join('/');
-}
+import { pathParts } from './utilities/path';
 
 export function applyChange(src: any, dest: any) {
   // if not mergeable, don't merge
